:recycle: Refactor into components
