+++ src/components/Footer.scss
@@ -1,0 +1,14 @@
+footer {
+  .copyright-text {
+    @apply mb-1;
+    @apply text-xs;
+    @apply text-center;
+  }
+
+  .no-affiliation-text {
+    @apply mb-1;
+    @apply text-xs;
+    @apply italic;
+    @apply text-center;
+  }
+}
