+++ src/components/Header.scss
@@ -1,0 +1,9 @@
+.page-title {
+  @apply text-5xl;
+  @apply mb-12;
+  font-family: 'Graduate', cursive;
+  @apply my-6;
+  @apply text-4xl;
+  @apply font-bold;
+  @apply text-center;
+}
